backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai==1.0.0
serpapi==0.1.5
python-dotenv==1.0.1
pydantic==2.9.2
